Moved AttributesComponents to separate namespace
