+++ src/GrasshopperTeklaDrawingLink/Components/AttributesComponents/BoltAttributesComponent.cs
@@ -3,7 +3,7 @@ using GTDrawingLink.Tools;
 using System.Drawing;
 using Tekla.Structures.Drawing;
 
-namespace GTDrawingLink.Components
+namespace GTDrawingLink.Components.AttributesComponents
 {
     public class BoltAttributesComponent : TeklaComponentBaseNew<BoltAttributesCommand>
     {
