+++ src/GrasshopperTeklaDrawingLink/Components/AttributesComponents/SymbolAttributesComponent.cs
@@ -5,7 +5,7 @@ using GTDrawingLink.Types;
 using System.Drawing;
 using Tekla.Structures.Drawing;
 
-namespace GTDrawingLink.Components
+namespace GTDrawingLink.Components.AttributesComponents
 {
     public class SymbolAttributesComponent : TeklaComponentBaseNew<SymbolAttributesCommand>
     {
